# Use cache for parsing files faster the second time