This should fix #876 (at least the ocaml part).
This saves parsed (and Naming_AST.resolved) generic AST
in /tmp/semgrep_core_cache<uid/ if the -use_parsing_cache command-line
argument was passed to semgrep-core.

To clean the cache, simply do
rm -f /tmp/semgrep_core_cache<uid/ from the Python side if you need to.

Test plan:
takes 2.1s the first time and 0.7sec the second time when testing
on tests/PERF/ js files.

/home/pad/github/semgrep/semgrep-core/_build/default/bin/Main.exe -equivalences /home/pad/github/semgrep/semgrep-core/data/basic_equivalences.yml -verbose -profile -e 1==2 -lang js tests/PERF/
---------------------
profiling result
---------------------
Main total                               :      1.996 sec          1 count
Parse_js.parse                           :      1.043 sec          4 count
Semgrep.check                            :      0.537 sec          4 count
Parse_js.tokens                          :      0.401 sec         13 count
Semgrep.match_e_e                        :      0.307 sec     132621 count
...

This time with -use_parsing_cache:
/home/pad/github/semgrep/semgrep-core/_build/default/bin/Main.exe -equivalences /home/pad/github/semgrep/semgrep-core/data/basic_equivalences.yml -use_parsing_cache -verbose -profile -e 1==2 -lang js tests/PERF/

---------------------
profiling result
---------------------
Main total                               :      0.546 sec          1 count
Semgrep.check                            :      0.486 sec          4 count
Semgrep.match_e_e                        :      0.285 sec     132621 count
rule:-e/-f                               :      0.105 sec     132621 count
Apply_equivalences.apply                 :      0.042 sec          4 count
Main.cache_computation                   :      0.040 sec          4 count
...

ls -al /tmp/semgrep_core_cache_1000/
total 9984
drwx------  2 pad  pad     4096 Jun  2 16:08 .
drwxrwxrwt 17 root root   49152 Jun  2 16:26 ..
-rw-r--r--  1 pad  pad  1019805 Jun  2 16:07 027e133a90b2c79d01c3224dfb0fed8f.ast_cache
-rw-r--r--  1 pad  pad  3953523 Jun  2 16:05 20d25efa622e4bfbc4015d03ba0d1d68.ast_cache
-rw-r--r--  1 pad  pad  3953523 Jun  2 16:08 6c05be109ca532cf545c082bda4a07fd.ast_cache
-rw-r--r--  1 pad  pad     7590 Jun  2 16:05 8cae6609c9048056880aa68cc383912c.ast_cache
-rw-r--r--  1 pad  pad     7590 Jun  2 16:07 92a4a1e679734b50b6dbc1cbfae37e86.ast_cache
-rw-r--r--  1 pad  pad    98103 Jun  2 16:07 a51fb3d6d3645f7004df0cee7b18e176.ast_cache
-rw-r--r--  1 pad  pad    98103 Jun  2 16:05 bd7d534ea53a44d782902098c97fedb9.ast_cache
-rw-r--r--  1 pad  pad  1019805 Jun  2 16:05 e763c7f5ef9424a2a7e295dd564cfabb.ast_cache

There are each time 2 files with the same size because I bumped ast_version
and so it regenerated new cache entries.