@mschwager 2020-06-02
Has there been any higher level discussion on the trade-offs between complexity and speed here? Caching can be simple at first glance, but difficult to get _completely_ correct, and can make debugging much more difficult. I.e. "why are we seeing these results? oh, there's some complex interaction with caching going on." <- this can be a pain getting right on the multitude of systems and environments we expect `semgrep` to run on.

For example, it seems there's already a lot of discussion around what the caching key should be. It looks like we currently think it should include:

* File path - can there be collisions here?
* Language, e.g. `--lang`
* `pfff` version
* A manually set AST version that we're asking developers to remember to bump at certain times

Is there anything we're missing? Are we painting ourselves in a corner right now with design decisions we may regret?

I'm also curious, what happens if `/tmp` doesn't exist? There are weird systems out there. This seems to depend on `Unix.getuid()` vs. something like `NamedTemporaryFile` in Python, which feels fragile. Are there collisions here we may not expect? How do we debug this functionality on an external user's system? Is it easier to implement this in Python vs. OCaml? E.g. hashing CLI args passed to `semgrep-core` + a hash of the file contents _should_ always produce the same results. Does this work with YAML URL functionality, or anything else that can change out from underneath us?

Basically, I'm trying to highlight the complexity of the solution, and implementing caching correctly in general. My vote would be to pass on this optimization until we see a greater need for it, or until we come up with a peer-reviewed design we believe to be rock-solid. We've already seen great speedups recently, and while 2.1s -> 0.7s is another great speedup, it doesn't warrant the additional complexity, IMO.

See also some discussion on similar functionality in `flake8`: https://gitlab.com/pycqa/flake8/-/issues/601