@brendongo 2020-06-02
@mschwager good flag re: complexity vs performance here. Definitely share your sentiments on how hard it is to get caching right. We've definitely been bit by caching bugs in bento. I'll hold of on merging this until #863 is ready so we can see what the performance hit that gets and what the performance gain of caching AST gives us.