@mjambon 2020-06-02
I agree with the concerns about getting caching right in general, but I'm not in a position to say whether this should be done here. It's good to have a flag to enable/disable caching, and there should be tests to check they produce the same results.