@mjambon 2020-06-02
Things that may have an impact: is the cached AST sensitive to command-line options or config files? (how many, is it going to change?) or worse, environment variables? Depending on how messy that is, caching could be a breeze or a nightmare.