@aryx 2020-06-02
Yes the solution here is small and easy to undo. I also would prefer no caching, but it can be useful to experiment and see how much it simplifies semgrep-python when you can call multiple times semgrepcore