@nbrahms 2020-06-02
As long as the caching isn't used by default, it makes sense to me to integrate it now so we can start playing with it.